+++ packages/integrity/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @meshql/core@0.3.0
+  - @meshql/http@0.2.1
+
 ## 0.2.0
 
 ### Minor Changes